Moved to numpy_compat
