+++ dask/array/__init__.py
@@ -23,6 +23,7 @@ try:
         to_zarr,
     )
     from .tiledb_io import from_tiledb, to_tiledb
+    from .numpy_compat import rollaxis, moveaxis
     from .routines import (
         take,
         choose,
@@ -79,8 +80,6 @@ try:
         fliplr,
         einsum,
         average,
-        rollaxis,
-        moveaxis,
     )
     from .reshape import reshape
     from .ufunc import (
